Move inverse Ops to sub-module
